+++ src/app/app.component.html
@@ -6,18 +6,6 @@
 </nav>
 -->
 
-<div style="text-align: center, bold; color: white" >
-  <h5>Cookies In Angular</h5>
-  <h5>Cookie Data:</h5>
-  <p>Name: {{cookie_name}}</p>
-  <p>Cookie Object: {{all_cookies | json}}</p>
-  <br>
-  <button (click)="setCookie()">Set Cookies</button>
-  <br><br>
-  <button (click)="deleteCookie()">Delete</button>
-  <br><br>
-  <button (click)="deleteAll()">Delete All</button>
-  </div>
 <!--bootstrap-->
 <header>
   <!-- Navbar -->
